# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoulForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{goalId}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by ID
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalWithSubtasks"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /goals/{goalId}/complete:
    post:
      operationId: completeGoal
      tags: [goals]
      summary: Mark goal as complete (awards XP)
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Completed goal with XP gained
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XPResult"

  # SUBTASKS
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List habits sorted by time of day
      responses:
        "200":
          description: Habits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{habitId}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /habits/{habitId}/checkin:
    post:
      operationId: checkinHabit
      tags: [habits]
      summary: Check in a habit for today (awards XP, damages boss)
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Checkin result with XP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XPResult"

  # BOSSES
...
  /bosses/{bossId}/challenge:
    post:
      operationId: challengeBoss
      tags: [bosses]
      summary: Start challenging a boss
      parameters:
        - name: bossId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Boss battle started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossBattle"

  # AI CHALLENGES
...
  /challenges/generate:
    post:
      operationId: generateChallenges
      tags: [challenges]
      summary: Generate AI challenges for a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeGenerateInput"
      responses:
        "201":
          description: Generated challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
  /challenges/{challengeId}:
    patch:
      operationId: updateChallenge
      tags: [challenges]
      summary: Update a challenge
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeUpdate"
      responses:
        "200":
          description: Updated challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
  /challenges/{challengeId}/complete:
    post:
      operationId: completeChallenge
      tags: [challenges]
      summary: Complete a challenge (awards bonus XP)
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: XP result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XPResult"

  # INVENTORY
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    Player:
      type: object
      required: [id, userId, username, level, xp, xpToNextLevel, hp, maxHp, gold, strength, endurance, dexterity, faith, avatarClass, equippedSkills, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
        level:
          type: integer
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        gold:
          type: integer
        strength:
          type: integer
        endurance:
          type: integer
        dexterity:
          type: integer
        faith:
          type: integer
        avatarClass:
          type: string
        equippedSkills:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    XPResult:
      type: object
      required: [xpGained, player, leveledUp, loot]
      properties:
        xpGained:
          type: integer
        player:
          $ref: "#/components/schemas/Player"
        leveledUp:
          type: boolean
        bossHpReduced:
          type: ["integer", "null"]
        loot:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"

    # GOALS
    Goal:
      type: object
      required: [id, userId, title, description, status, priority, xpReward, deadline, createdAt, completedAt, subtaskCount, completedSubtaskCount]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, abandoned]
        priority:
          type: string
          enum: [low, medium, high, legendary]
        xpReward:
          type: integer
        deadline:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        subtaskCount:
          type: integer
        completedSubtaskCount:
          type: integer
    GoalWithSubtasks:
      type: object
      required: [id, userId, title, description, status, priority, xpReward, deadline, createdAt, completedAt, subtasks]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, abandoned]
        priority:
          type: string
          enum: [low, medium, high, legendary]
        xpReward:
          type: integer
        deadline:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"
    GoalInput:
      type: object
      required: [title, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, legendary]
        deadline:
          type: string
        xpReward:
          type: integer
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, legendary]
        deadline:
          type: string
        status:
          type: string
          enum: [active, completed, abandoned]

    # SUBTASKS
    Subtask:
      type: object
      required: [id, goalId, userId, title, completed, xpReward, createdAt, completedAt]
      properties:
        id:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        completed:
          type: boolean
        xpReward:
          type: integer
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
...
    Habit:
      type: object
      required: [id, userId, title, description, frequency, timeOfDay, xpReward, streak, lastCheckin, createdAt, completedToday]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: ["string", "null"]
        frequency:
          type: string
          enum: [daily, weekly]
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, anytime]
        xpReward:
          type: integer
        streak:
          type: integer
        lastCheckin:
          type: ["string", "null"]
        createdAt:
          type: string
        completedToday:
          type: boolean
    HabitInput:
      type: object
      required: [title, frequency, timeOfDay]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly]
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, anytime]
        xpReward:
          type: integer
    HabitUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly]
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, anytime]
        xpReward:
          type: integer

    # BOSSES
    Boss:
      type: object
      required: [id, name, description, difficulty, maxHp, xpReward, goldReward, lootTable, imageUrl, isCustom, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        maxHp:
          type: integer
        xpReward:
          type: integer
        goldReward:
          type: integer
        lootTable:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        isCustom:
          type: boolean
        createdAt:
          type: string
...
    BossBattle:
      type: object
      required: [id, userId, boss, currentHp, startedAt, attackPoints]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        boss:
          $ref: "#/components/schemas/Boss"
        currentHp:
          type: integer
        startedAt:
          type: string
        attackPoints:
          type: integer
...
    Challenge:
      type: object
      required: [id, userId, goalId, title, description, duration, difficulty, xpReward, status, createdAt, completedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        goalId:
          type: ["string", "null"]
          format: uuid
        title:
          type: string
        description:
          type: string
        duration:
          type: string
          enum: [daily, weekly, monthly]
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        xpReward:
          type: integer
        status:
          type: string
          enum: [active, completed, abandoned]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    ChallengeGenerateInput:
      type: object
      required: [goalId, duration]
      properties:
        goalId:
          type: string
          format: uuid
        duration:
          type: string
          enum: [daily, weekly, monthly]
        count:
          type: integer
    ChallengeUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, abandoned]

    # INVENTORY
    InventoryItem:
      type: object
      required: [id, userId, name, description, rarity, type, obtainedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        type:
          type: string
          enum: [weapon, armor, accessory, consumable, trophy]
        obtainedAt:
          type: string

    # SKILLS
...