# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoulForge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeDifficulty } from './challengeDifficulty';
import type { ChallengeDuration } from './challengeDuration';
import type { ChallengeStatus } from './challengeStatus';

export interface Challenge {
  id: string;
  userId: string;
  /** @nullable */
  goalId: string | null;
  title: string;
  description: string;
  duration: ChallengeDuration;
  difficulty: ChallengeDifficulty;
  xpReward: number;
  status: ChallengeStatus;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
}